python/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azurepg-entra"
version = "0.0.1"
description = "Enable persistent connections to Azure PostgreSQL via Entra ID and automated token refresh"
readme = "README.md"
authors = [
  {name = "Arjun Narendra", email = "v-anarendra@microsoft.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "azure-identity>=1.13.0",
  "azure-core>=1.24.0",
  "aiohttp>=3.8.0",
  "PyJWT>=2.0.0"
]

[project.optional-dependencies]
# psycopg3 support
psycopg3 = [
  "psycopg[binary]>=3.1.0",
]

# psycopg2 support
psycopg2 = [
  "psycopg2-binary>=2.9.0"
]

# SQLAlchemy support
sqlalchemy = [
  "sqlalchemy>=2.0.0",
]

# Development dependencies
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "python-dotenv>=1.0.0",
  "mypy ~= 1.15"
]

# All optional dependencies combined
all = [
  "psycopg[binary]>=3.1.0",
  "psycopg-pool>=3.1.0",
  "psycopg2-binary>=2.9.0",
  "aiopg>=1.4.0", 
  "sqlalchemy>=2.0.0",
  "asyncpg>=0.28.0",
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "python-dotenv>=1.0.0",
  "mypy ~= 1.15"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["azurepg_entra*"]

[project.urls]
Homepage = "https://github.com/v-anarendra_microsoft/entra-id-integration-for-drivers"
Issues = "https://github.com/v-anarendra_microsoft/entra-id-integration-for-drivers/issues"

# Installation examples:
# Basic package only:                        pip install -e .
# With psycopg3 support:                     pip install -e ".[psycopg3]"
# With psycopg2 support:                     pip install -e ".[psycopg2]" 
# With SQLAlchemy support:                   pip install -e ".[sqlalchemy]"
# With development tools:                    pip install -e ".[dev]"
# Everything:                                pip install -e ".[all]"